feat: 實作活動紀錄 API 並進行全面優化
新增活動紀錄完整 CRUD 操作，包含 7 個 REST 端點

實作 LIVE / MANUAL 紀錄類型的商業邏輯與驗證

建立自訂例外階層（BadRequest、NotFound、Conflict），並統一錯誤回應格式

優化資料庫查詢：修正 JPQL 相容性，使用 PostgreSQL 特定函數

將 sum 查詢回傳型別由 Integer 改為 Long，避免溢位

熱門活動查詢由 Page<UUID> 改為 List<UUID>（移除不必要的 count）

將 Between 方法改為半開區間 [from, to) 查詢

新增完整的篩選與分頁支援

確保使用者歸屬檢查與身份驗證安全

保持向下相容性（舊方法標記為 Deprecated）

API 端點：

POST /api/records — 建立紀錄

PATCH /api/records/{id}/finish — 結束 LIVE 紀錄

PUT /api/records/{id} — 更新紀錄

DELETE /api/records/{id} — 刪除紀錄

GET /api/records/{id} — 取得單一紀錄

GET /api/records — 查詢紀錄（支援篩選）

GET /api/records/running — 查詢進行中的紀錄

重大變更 (Breaking Changes)： 無（保持相容性）
效能 (Performance)： 查詢大幅優化，減少資料庫呼叫
安全性 (Security)： 完整的使用者歸屬檢查與身份驗證

+++ backend/mymomentum/src/main/java/com/ramble/mymomentum/service/ActivityService.java
@@ -192,8 +192,8 @@ public class ActivityService {
      */
     private ActivityResponse convertToActivityResponse(Activity activity) {
         // Calculate total time (in minutes)
-        Integer totalTimeSeconds = activityRecordRepository.getTotalTimeByActivityId(activity.getId());
-        Integer totalTimeMinutes = totalTimeSeconds / 60;
+        Long totalTimeSeconds = activityRecordRepository.getTotalTimeByActivityId(activity.getId());
+        Long totalTimeMinutes = totalTimeSeconds / 60;
 
         // Calculate weekly time (in minutes) - from start of current week
         Instant weekStart = Instant.now()
@@ -201,20 +201,20 @@ public class ActivityService {
                 .truncatedTo(ChronoUnit.DAYS)
                 .with(TemporalAdjusters.previousOrSame(java.time.DayOfWeek.MONDAY))
                 .toInstant();
-        Integer weeklyTimeSeconds = activityRecordRepository.getWeeklyTimeByActivityId(activity.getId(), weekStart);
-        Integer weeklyTimeMinutes = weeklyTimeSeconds / 60;
+        Long weeklyTimeSeconds = activityRecordRepository.getWeeklyTimeByActivityId(activity.getId(), weekStart);
+        Long weeklyTimeMinutes = weeklyTimeSeconds / 60;
 
         // Convert target time from seconds to minutes
         Integer targetTimeMinutes = activity.getTargetTime() != null ? activity.getTargetTime() / 60 : 0;
 
-        return new ActivityResponse(
-            activity.getId().toString(),
-            activity.getName(),
-            totalTimeMinutes,
-            weeklyTimeMinutes,
-            targetTimeMinutes,
-            activity.getColor(),
-            activity.getIcon()
+                return new ActivityResponse(
+                activity.getId().toString(),
+                activity.getName(),
+                totalTimeMinutes.intValue(),
+                weeklyTimeMinutes.intValue(),
+                targetTimeMinutes,
+                activity.getColor(),
+                activity.getIcon()
         );
     }
 }
